+++ workspaces/scorecard/playwright.config.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Backstage Authors
+ * Copyright Red Hat, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,44 +17,33 @@
 import { defineConfig } from '@playwright/test';
 import { generateProjects } from '@backstage/e2e-test-utils/playwright';
 
-/**
- * See https://playwright.dev/docs/test-configuration.
- */
 export default defineConfig({
-  timeout: 60_000,
+  timeout: 2 * 60 * 1000,
 
   expect: {
-    timeout: 5_000,
+    timeout: 5000,
   },
 
-  // Run your local dev server before starting the tests
-  webServer: process.env.CI
+  webServer: process.env.PLAYWRIGHT_URL
     ? []
-    : [
-        {
-          command: 'yarn start',
-          port: 3000,
-          reuseExistingServer: true,
-          timeout: 60_000,
-        },
-      ],
-
-  forbidOnly: !!process.env.CI,
+    : {
+        command: 'yarn start',
+        port: 3000,
+        reuseExistingServer: false,
+      },
 
   retries: process.env.CI ? 2 : 0,
 
   reporter: [['html', { open: 'never', outputFolder: 'e2e-test-report' }]],
 
   use: {
+    baseURL: process.env.PLAYWRIGHT_URL ?? 'http://localhost:3000',
     actionTimeout: 0,
-    baseURL:
-      process.env.PLAYWRIGHT_URL ??
-      (process.env.CI ? 'http://localhost:7007' : 'http://localhost:3000'),
     screenshot: 'only-on-failure',
     trace: 'on-first-retry',
   },
 
   outputDir: 'node_modules/.cache/e2e-test-results',
 
-  projects: generateProjects(), // Find all packages with e2e-test folders
+  projects: generateProjects(), // Auto-discover e2e-tests in packages
 });