+++ workspaces/global-header/plugins/global-header/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @red-hat-developer-hub/backstage-plugin-global-header
 
+## 1.17.2
+
+### Patch Changes
+
+- 4e2f33e: Fix the test package build and rebuild both packages to have a consistant commit for them.
+- 2c4b3c0: Updated dependency `@red-hat-developer-hub/backstage-plugin-theme` to `^0.10.0`.
+
 ## 1.17.1
 
 ### Patch Changes